Time: 0 ms (100%), Space: 11.7 MB (38.86%) - LeetHub

+++ 1636-number-of-substrings-with-only-1s/1636-number-of-substrings-with-only-1s.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int numSub(string s) {
         long long count = 0, length = 0;
-        const long long mod = 1000000007;
+        const long mod = 1000000007;
 
         for (int i = 0; i < s.length(); ++i) {
             if (s[i] == '1') {
